rename and rearrange files and tests in where package
